add attention  lstm rnn gru and dataloader and so on and push the javacpp-pytorch version to 2.1-1.15.1 all

+++ examples/src/main/scala/gpt/Utils.scala
@@ -213,7 +213,7 @@ object Utils:
       // get the device
       val cudaAllocator = torch_cuda.getAllocator()
       println(cudaAllocator.initialized())
-      val stats = cudaAllocator.getDeviceStats(device)
+      val stats = cudaAllocator.getDeviceStats(device.toByte)
 
       // Collect the statistics
       val result = scala.collection.mutable.Map[String, Long]()
